clickpm/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-filter==23.1
django-cors-headers==3.14.0
psycopg2-binary==2.9.6
python-decouple==3.8
Pillow==9.5.0
setuptools
gunicorn==21.2.0
whitenoise==6.6.0

